Updated fiiles

+++ build.ps1
@@ -15,9 +15,11 @@ if (Test-Path -Path "$PSScriptRoot/obj") {
 
 # Copy private keys to access git repo
 if (-not (Test-Path -Path "$PSScriptRoot/docker/id_rsa")) {
-    if ($env:GIT_PRIVATE_KEY -ne $null) {
+    if (-not [string]::IsNullOrEmpty($env:GIT_PRIVATE_KEY)) {
+        Write-Host "Creating docker/id_rsa from environment variable..."
         Set-Content -Path "$PSScriptRoot/docker/id_rsa" -Value $env:GIT_PRIVATE_KEY
     } elseif (Test-Path -Path "~/.ssh/id_rsa") {
+        Write-Host "Copying ~/.ssh/id_rsa to docker..."
         Copy-Item -Path "~/.ssh/id_rsa" -Destination "docker"
     } else {
         Write-Host "Missing ~/.ssh/id_rsa file..."
@@ -26,12 +28,17 @@ if (-not (Test-Path -Path "$PSScriptRoot/docker/id_rsa")) {
 }
 
 # Copy .npmrc to docker folder to use it inside container
-if (Test-Path -Path "~/.npmrc") {
-    Write-Host "Putting ~/.npmrc to docker folder..."
-    Copy-Item -Path "~/.npmrc" -Destination "docker" 
-} else {
-    Write-Host "Missing ~/.npmrc file..."
-    Set-Content -Path "$PSScriptRoot/docker/.npmrc" -Value ""
+if (-not (Test-Path -Path "$PSScriptRoot/docker/.npmrc")) {
+    if (-not [string]::IsNullOrEmpty($env:NPM_TOKEN)) {
+        Write-Host "Creating docker/.npmrc from environment variable..."
+        Set-Content -Path "$PSScriptRoot/docker/.npmrc" -Value "//registry.npmjs.org/:_authToken=$($env:NPM_TOKEN)"
+    } elseif (Test-Path -Path "~/.npmrc") {
+        Write-Host "Copying ~/.npmrc to docker..."
+        Copy-Item -Path "~/.npmrc" -Destination "docker" 
+    } else {
+        Write-Host "Missing ~/.npmrc file..."
+        Set-Content -Path "$PSScriptRoot/docker/.npmrc" -Value ""
+    }
 }
 
 # Build docker image
